Create URL Sanitization
Something I made when I needed to deal with Fortify Open Redirect Issues on some code I was working on.  I didn't find a whole lot out there and what I found was more/less MVC and/or it was very minimal.  This works on both full URLs (containing http) as well as just page paths, and support querystrings.  One thing is you will need a list to store in valid page names and a list to store in valid querystring parameters.  If you just have pages and no querystrings then you can comment out the "areParmsValid" bool check and be sure to remove it from the final if statement checking that bool.